refactor(plugins): drop dead plugins, merge voice adapters, generic bundle prune
