+++ apps/host/src/agent/application/sessionSource.ts
@@ -20,7 +20,6 @@ import type {
     SessionStatus,
     StreamingBehavior,
     PluginsInvalidatedFrame,
-    VoiceProviderOption,
     WatchSettlement,
 } from '@muxr/contract';
 
@@ -113,8 +112,6 @@ export interface SessionSource {
     /** Declared RPC mode for a catalog contribution, so read-only devices can be allowed through read paths only. */
     pluginRpcMode?(options: { pluginId: string; manifestHash: string; contributionId: string }): 'read' | 'write' | undefined;
     pluginStream(options: { deviceId: string; pluginId: string; manifestHash: string; contributionId: string; channel: string; sessionId?: string }): Promise<null>;
-    voiceProviderList(): Promise<VoiceProviderOption[]>;
-    voiceProviderSelect(providerId: string): Promise<VoiceProviderOption[]>;
     /** Split layout of one tab (rects in terminal cells) for grid views. */
     herdrLayout(tabId: string): Promise<{
         tabId: string;
