+++ apps/host/src/agent/infrastructure/herdrSessionSource.ts
@@ -28,7 +28,6 @@ import type {
     SessionSnapshot,
     SessionStartResult,
     SessionStatus,
-    VoiceProviderOption,
 } from '@muxr/contract';
 import { ATTENTION_REASONS, parseCloseResult, realtimePluginPublicContext, relayControlUrl } from '@muxr/contract';
 import { AttachmentWatcher } from './attachmentWatcher.js';
@@ -1813,66 +1812,6 @@ export async function createHerdrSessionSource(
         await reconcilePlugins(true);
     }
 
-    function voiceProviderOptions(): VoiceProviderOption[] {
-        return catalog.capabilityPlugins('voice.session').map(({ pluginId, name, enabled, source, hasBackend }) => ({
-            id: pluginId,
-            name,
-            selected: enabled,
-            source,
-            hasBackend,
-        }));
-    }
-
-    async function selectVoiceProviderNow(providerId: string): Promise<VoiceProviderOption[]> {
-        await refreshPlugins();
-        const providers = catalog.capabilityPlugins('voice.session');
-        const target = providers.find((candidate) => candidate.pluginId === providerId);
-        if (target === undefined) throw new Error('realtime voice provider is not installed on this machine; update muxr first');
-        const previouslyEnabled = providers.filter(({ enabled }) => enabled);
-        const disabled: typeof previouslyEnabled = [];
-        let enabledTarget = false;
-        try {
-            for (const current of previouslyEnabled) {
-                if (current.pluginId === target.pluginId) continue;
-                await client.call('plugin.disable', { plugin_id: current.pluginId });
-                disabled.push(current);
-            }
-            if (!target.enabled) {
-                await client.call('plugin.enable', { plugin_id: target.pluginId });
-                enabledTarget = true;
-            }
-            await refreshPlugins();
-            const latest = catalog.capabilityPlugins('voice.session');
-            let converged = false;
-            if (!latest.some((candidate) => candidate.pluginId === target.pluginId && candidate.enabled)) {
-                await client.call('plugin.enable', { plugin_id: target.pluginId });
-                enabledTarget = true;
-                converged = true;
-            }
-            for (const current of latest) {
-                if (!current.enabled || current.pluginId === target.pluginId) continue;
-                await client.call('plugin.disable', { plugin_id: current.pluginId });
-                converged = true;
-            }
-            if (converged) await refreshPlugins();
-            return voiceProviderOptions();
-        } catch (error) {
-            if (enabledTarget) await client.call('plugin.disable', { plugin_id: target.pluginId }).catch(() => undefined);
-            for (const current of disabled.reverse()) {
-                await client.call('plugin.enable', { plugin_id: current.pluginId }).catch(() => undefined);
-            }
-            await refreshPlugins().catch(() => undefined);
-            throw error;
-        }
-    }
-
-    let voiceSwitch: Promise<void> = Promise.resolve();
-    function selectVoiceProvider(providerId: string): Promise<VoiceProviderOption[]> {
-        const run = voiceSwitch.then(() => selectVoiceProviderNow(providerId));
-        voiceSwitch = run.then(() => undefined, () => undefined);
-        return run;
-    }
-
     void reconcilePlugins().catch(() => undefined);
     pluginPollTimer = setInterval(() => {
         if (machineListeners.size > 0) void reconcilePlugins().catch(() => undefined);
@@ -2129,15 +2068,6 @@ export async function createHerdrSessionSource(
             await refreshPlugins();
             if (approved) catalog.manifest(pluginId, manifestHash);
             await pluginApprovals.set(deviceId, pluginId, approved);
-        },
-
-        async voiceProviderList() {
-            await refreshPlugins();
-            return voiceProviderOptions();
-        },
-
-        async voiceProviderSelect(provider) {
-            return selectVoiceProvider(provider);
         },
 
         async pluginInvoke({ deviceId, pluginId, manifestHash, contributionId, sessionId, idempotencyKey }) {
