+++ apps/host/src/requests/application/createRequestDispatcher.ts
@@ -61,7 +61,7 @@ type PluginExecutionRequest = Extract<ClientRequest, {
 
 const VIEW_ONLY_REQUESTS: ReadonlySet<RequestType> = new Set([
     'session.list', 'session.open', 'session.status',
-    'herdr.tree', 'herdr.agentKinds', 'herdr.layout', 'pane.read', 'plugin.list', 'plugin.manifest', 'voice.provider.list',
+    'herdr.tree', 'herdr.agentKinds', 'herdr.layout', 'pane.read', 'plugin.list', 'plugin.manifest',
     'attachment.fetch', 'attachment.read', 'unread.catalog',
     'attention.catalog', 'lifecycle.catalog', 'machines.list', 'terminal.attach',
 ]);
@@ -138,8 +138,6 @@ export function createRequestDispatcher(options: RequestDispatcherOptions): {
         'plugin.invoke': () => { throw new Error('authenticated device context required'); },
         'plugin.call': () => { throw new Error('authenticated device context required'); },
         'plugin.stream': () => { throw new Error('authenticated device context required'); },
-        'voice.provider.list': () => source.voiceProviderList(),
-        'voice.provider.select': (params) => source.voiceProviderSelect(params.providerId),
         'herdr.cli': async (params) => {
             const result = await runHerdrCli(params.args, params.timeoutMs);
             await source.refreshHerdr();
