--- plugins/inbox/herdr-plugin.toml
@@ -1,6 +1,0 @@
-id = "muxr.inbox"
-name = "Inbox"
-version = "0.1.0"
-min_herdr_version = "0.8.0"
-platforms = ["linux", "macos"]
-description = "Grouped session collection built from bounded public host context"
