--- plugins/servers/README.md
@@ -1,7 +1,0 @@
-# Servers
-
-What is listening on this machine, and project dev servers for Preview.
-
-- **Ports:** every listener with process and pid; stop stray processes from your phone (declared write).
-- **Preview:** detect HTML frontends already running in the selected project and open them through the paired relay.
-- **Removal:** `herdr plugin unlink muxr.servers`.
