--- plugins/servers/muxr-ui.json
@@ -1,127 +1,0 @@
-{
-  "schemaVersion": 1,
-  "pluginId": "muxr.servers",
-  "minMuxrVersion": 5,
-  "contributions": [
-    {
-      "slot": "host.rpc",
-      "id": "ports.list",
-      "type": "rpc",
-      "method": "list",
-      "entry": "ports.mjs",
-      "mode": "read"
-    },
-    {
-      "slot": "host.rpc",
-      "id": "ports.detail",
-      "type": "rpc",
-      "method": "detail",
-      "entry": "ports.mjs",
-      "mode": "read"
-    },
-    {
-      "slot": "host.rpc",
-      "id": "ports.stop",
-      "type": "rpc",
-      "method": "stop",
-      "entry": "ports.mjs",
-      "mode": "write"
-    },
-    {
-      "slot": "navigation.primary",
-      "id": "ports.nav",
-      "type": "navigation-item",
-      "label": "Ports",
-      "icon": "git-network-outline",
-      "contentContributionId": "ports.browse"
-    },
-    {
-      "slot": "navigation.content",
-      "id": "ports.browse",
-      "type": "screen",
-      "title": "{{data.title}}",
-      "data": {
-        "type": "plugin.call",
-        "contributionId": "ports.list"
-      },
-      "children": [
-        {
-          "type": "list",
-          "title": "Listening",
-          "emptyText": "Nothing listening",
-          "rows": [],
-          "repeat": {
-            "path": "data.ports",
-            "template": {
-              "type": "row",
-              "title": "{{item.title}}",
-              "subtitle": "{{item.subtitle}}",
-              "action": {
-                "type": "screen",
-                "contributionId": "ports.port",
-                "params": {
-                  "pid": "{{item.pid}}"
-                }
-              }
-            }
-          }
-        }
-      ]
-    },
-    {
-      "slot": "navigation.content",
-      "id": "ports.port",
-      "type": "screen",
-      "title": "{{data.title}}",
-      "data": {
-        "type": "plugin.call",
-        "contributionId": "ports.detail"
-      },
-      "children": [
-        {
-          "type": "metric",
-          "label": "Process",
-          "value": "{{data.name}}"
-        },
-        {
-          "type": "metric",
-          "label": "PID",
-          "value": "{{data.pid}}"
-        },
-        {
-          "type": "button",
-          "label": "Stop this process",
-          "action": {
-            "type": "plugin.call",
-            "contributionId": "ports.stop"
-          },
-          "variant": "danger"
-        }
-      ]
-    },
-    {
-      "slot": "session.header.trailing",
-      "id": "listeners",
-      "type": "native",
-      "primitive": "item-list",
-      "params": {
-        "source": {
-          "type": "plugin.call",
-          "contributionId": "servers.listeners"
-        },
-        "title": "Preview",
-        "icon": "globe-outline",
-        "accessibilityLabel": "Open project server",
-        "refreshIntervalMs": 15000
-      }
-    },
-    {
-      "slot": "host.rpc",
-      "id": "servers.listeners",
-      "type": "rpc",
-      "method": "list",
-      "entry": "serve.mjs",
-      "mode": "read"
-    }
-  ]
-}
