--- plugins/voice-gemini/muxr-ui.json
@@ -1,189 +1,0 @@
-{
-  "schemaVersion": 1,
-  "pluginId": "muxr.voice-gemini",
-  "minMuxrVersion": 6,
-  "capabilities": {
-    "voice.status": "status",
-    "voice.key.set": "key-set",
-    "voice.session": "session",
-    "voice.report": "report"
-  },
-  "contributions": [
-    {
-      "slot": "host.rpc",
-      "id": "status",
-      "type": "rpc",
-      "method": "status",
-      "entry": "rpc.mjs",
-      "mode": "read"
-    },
-    {
-      "slot": "host.rpc",
-      "id": "key-set",
-      "type": "rpc",
-      "method": "key.set",
-      "entry": "rpc.mjs",
-      "mode": "write"
-    },
-    {
-      "slot": "host.rpc",
-      "id": "key-clear",
-      "type": "rpc",
-      "method": "key.clear",
-      "entry": "rpc.mjs",
-      "mode": "write"
-    },
-    {
-      "slot": "host.stream",
-      "id": "session",
-      "type": "stream",
-      "entry": "stream.mjs"
-    },
-    {
-      "slot": "host.rpc",
-      "id": "report",
-      "type": "rpc",
-      "method": "report",
-      "entry": "rpc.mjs",
-      "mode": "read"
-    },
-    {
-      "slot": "app.overlay",
-      "id": "overlay",
-      "type": "native",
-      "primitive": "realtime-session-overlay"
-    },
-    {
-      "slot": "home.composer.leading",
-      "id": "talk-home",
-      "type": "native",
-      "primitive": "icon-button",
-      "params": {
-        "capability": "voice.start",
-        "icon": "pulse-outline",
-        "accessibilityLabel": "Start or open realtime conversation",
-        "indicator": "realtime-session"
-      }
-    },
-    {
-      "slot": "session.composer.trailing",
-      "id": "talk-session",
-      "type": "native",
-      "primitive": "icon-button",
-      "params": {
-        "capability": "voice.start",
-        "icon": "pulse-outline",
-        "accessibilityLabel": "Talk to this session",
-        "indicator": "realtime-session"
-      }
-    },
-    {
-      "slot": "settings.items",
-      "id": "settings",
-      "type": "settings-item",
-      "label": {
-        "default": "Gemini Live",
-        "translations": {
-          "ru": "Голосовой ассистент",
-          "pl": "Asystent głosowy",
-          "es": "Asistente de voz",
-          "it": "Assistente vocale",
-          "pt": "Assistente de voz",
-          "ca": "Assistent de veu",
-          "zh-Hans": "语音助手",
-          "zh-Hant": "語音助理",
-          "ja": "音声アシスタント"
-        }
-      },
-      "subtitle": {
-        "default": "Configure Gemini on the connected machine",
-        "translations": {
-          "ru": "Настройка предпочтений голосового взаимодействия",
-          "pl": "Konfiguruj preferencje interakcji głosowej",
-          "es": "Configura las preferencias de voz",
-          "it": "Configura le preferenze vocali",
-          "pt": "Configure as preferências de interação por voz",
-          "ca": "Configura les preferències d'interacció per veu",
-          "zh-Hans": "配置语音交互偏好",
-          "zh-Hant": "設定語音互動偏好",
-          "ja": "音声操作の設定"
-        }
-      },
-      "icon": "pulse-outline",
-      "action": {
-        "type": "screen",
-        "contributionId": "settings-screen"
-      }
-    },
-    {
-      "slot": "navigation.content",
-      "id": "settings-screen",
-      "type": "screen",
-      "title": "Gemini Live",
-      "data": {
-        "type": "plugin.call",
-        "contributionId": "status"
-      },
-      "children": [
-        {
-          "type": "text",
-          "text": "The Gemini key is sent directly to the connected machine and is never stored or displayed by the phone."
-        },
-        {
-          "type": "row",
-          "title": "Gemini key",
-          "value": "{{data.statusLabel}}"
-        },
-        {
-          "type": "button",
-          "label": "Set or replace Gemini key",
-          "variant": "primary",
-          "action": {
-            "type": "secure-prompt",
-            "title": "Gemini API key",
-            "message": "Sent once to this machine and stored only there.",
-            "placeholder": "AIza…",
-            "inputKey": "key",
-            "submit": {
-              "type": "plugin.call",
-              "contributionId": "key-set"
-            }
-          }
-        },
-        {
-          "type": "button",
-          "label": "Clear Gemini key",
-          "variant": "danger",
-          "action": {
-            "type": "confirm",
-            "title": "Clear Gemini key?",
-            "message": "Realtime voice will stop until a new key is configured.",
-            "confirmLabel": "Clear",
-            "destructive": true,
-            "action": {
-              "type": "plugin.call",
-              "contributionId": "key-clear"
-            }
-          }
-        }
-      ]
-    },
-    {
-      "slot": "events",
-      "id": "report-when-agent-stops",
-      "on": "agent.status",
-      "from": "working",
-      "to": [
-        "idle",
-        "done",
-        "blocked",
-        "failed"
-      ],
-      "action": {
-        "type": "capability",
-        "name": "speech.wake",
-        "include": "pane"
-      }
-    }
-  ]
-}
