+++ scripts/plugin/domain/index.ts
@@ -3,7 +3,5 @@ export {
     isPluginId,
     parseBundledPlugin,
     parsePluginId,
-    retiredSuccessor,
     type BundledPlugin,
-    type RetiredBundled,
 } from './pluginId.js';
